# pages/swagger-config.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Xavier Api",
    "description": "Welcome to the Xavier Api, your gateway to streamlined and efficient application integration. Designed with simplicity and versatility in mind, Xavier Api empowers developers to effortlessly manage and interact with essential data and functionalities. Whether you're building a web application, mobile app, or integrating with other services, Laurine Api provides a robust foundation for seamless operations.",
    "contact": {
      "email": "https://t.me/FVendettaV"
    },
    "license": {
      "name": "API Development Information",
      "url": "https://whatsapp.com/channel/0029VbACugX6rsQyHw0Wvr1g"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://japi.vercel.app",
      "description": "Xavier Api"
    }
  ],
  "paths": {
...
    "/api/stalk/ffstalk": {
      "post": {
        "tags": ["Stalk"],
        "description": "The ffStalk function is used to retrieve basic information about a Free Fire user (such as the username) from the Dunia Games top-up service. It sends the userId to an API and receives a response containing player details, such as the username, which is then returned as an object.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "default": "1921575755"
                  }
                },
                "required": ["id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/stalk/mlstalk": {
      "get": {
        "tags": ["Stalk"],
        "description": "mlStalk is a function designed to retrieve basic information about a Mobile Legends user. It works by sending a userId to a specific API endpoint, which then returns details about the user, such as their username or other relevant data. The function processes the API response and returns the user information in an object format.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/search/youtubesearch": {
      "get": {
        "tags": ["Search"],
        "description": "YT-Search or YouTubeSearch is a JavaScript library used to search for videos on YouTube directly through code without requiring the official YouTube API. This library is commonly used in the development of bots or applications that need a YouTube video search feature without obtaining an API key from Google.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/search/wikipedia": {
      "get": {
        "tags": ["Search"],
        "description": "Wikipedia is a free online encyclopedia that anyone can edit. Launched in 2001 by Jimmy Wales and Larry Sanger, it is managed by the Wikimedia Foundation and is available in over 300 languages. Wikipedia relies on volunteer contributions to create its articles and remains a major source of information on the internet.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/tools/detectporn": {
      "get": {
        "tags": ["Tools"],
        "description": "Porn detection is the process of identifying images or videos containing adult content using AI and Computer Vision. Common techniques include CNN (Convolutional Neural Networks), skin detection, and pre-trained models like NSFWJS, DeepAI NSFW API, or Google Cloud Vision. It is used for content moderation and digital safety.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/tools/subdofinder": {
      "get": {
        "tags": ["Tools"],
        "description": "SubdoFinder is a tool for discovering subdomains of a main domain, used in pentesting and cybersecurity. It utilizes brute-force, passive enumeration, and active scanning to identify hidden subdomains that may pose security risks.",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/tts/elevenlabs": {
      "get": {
        "tags": ["Text To Speech"],
        "description": "ElevenLabs is a technology company that develops AI-powered text-to-speech (TTS) solutions, known for their highly realistic and natural-sounding voices that can mimic human accents, intonations, and emotions, making them ideal for applications such as audiobooks, dubbing, narration, virtual assistants, gaming, and film production, while also supporting multiple languages and offering voice replication technology for content creators and developers.",
        "parameters": [
          {
            "name": "text",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
              }
           },
           {
            "name": "apiKey",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
              }
           },
           {
            "name": "voiceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
              }
           }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}